Move WebViewActivity to where it should be
