+++ app/src/main/AndroidManifest.xml
@@ -58,7 +58,7 @@
         <activity
             android:name=".ui.reader.ReaderActivity" />
         <activity
-            android:name=".ui.manga.info.WebViewActivity"
+            android:name=".ui.webview.WebViewActivity"
             android:configChanges="uiMode|orientation|screenSize"/>
         <activity
             android:name=".ui.main.BiometricActivity" />
